# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication and Authorization API",
        "title": "Auth API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/pdf/rotate": {
            "post": {
                "description": "PDF fayl sahifalarini aylantiradi (90, 180 yoki 270 gradus)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF Edit"
                ],
                "summary": "Rotate PDF pages",
                "parameters": [
                    {
                        "description": "Rotate request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RotatePDFRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/rotate/{id}": {
            "get": {
                "description": "Rotate PDF jarayonining natijasini olish",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF Edit"
                ],
                "summary": "Get Rotate Job info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RotateJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/api/pdf/split": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-split"
                ],
                "summary": "Create split job",
                "parameters": [
                    {
                        "description": "split job",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateSplitJobRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/split/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pdf-split"
                ],
                "summary": "Get split job by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "split job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SplitJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/unlock": {
            "post": {
                "description": "Qulflangan PDF faylni ochish (parolsiz qilish)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF Unlock"
                ],
                "summary": "Unlock PDF file",
                "parameters": [
                    {
                        "description": "Unlock PDF request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UnlockPDFRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/pdf/unlock/{id}": {
            "get": {
                "description": "Unlock PDF ish holatini ko‘rish",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PDF Unlock"
                ],
                "summary": "Get Unlock Job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UnlockPDFJob"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateSplitJobRequest": {
            "type": "object",
            "required": [
                "input_file_id",
                "split_ranges"
            ],
            "properties": {
                "input_file_id": {
                    "type": "string"
                },
                "split_ranges": {
                    "description": "Masalan: \"1-3,4-5\"",
                    "type": "string"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "description": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "models.RotateJob": {
            "type": "object",
            "properties": {
                "angle": {
                    "description": "Burilish burchagi",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_id": {
                    "type": "string"
                },
                "output_path": {
                    "type": "string"
                },
                "pages": {
                    "description": "Tanlangan sahifalar",
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.RotatePDFRequest": {
            "type": "object",
            "required": [
                "angle",
                "input_file_id",
                "pages"
            ],
            "properties": {
                "angle": {
                    "description": "Burchak: 90, 180, 270",
                    "type": "integer"
                },
                "input_file_id": {
                    "type": "string"
                },
                "pages": {
                    "description": "Sahifa raqamlari masalan: \"1-3\", \"2\", \"odd\"",
                    "type": "string"
                }
            }
        },
...
        "models.SplitJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_file_id": {
                    "type": "string"
                },
                "output_file_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "split_ranges": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "description": "❗️ string emas, *string bo‘lishi kerak",
                    "type": "string"
                }
            }
        },
        "models.UnlockPDFJob": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Yaratilgan vaqti",
                    "type": "string"
                },
                "id": {
                    "description": "Job ID",
                    "type": "string"
                },
                "input_file_id": {
                    "description": "Kiruvchi fayl IDsi",
                    "type": "string"
                },
                "output_file_id": {
                    "description": "Natijaviy fayl IDsi",
                    "type": "string"
                },
                "status": {
                    "description": "pending, done, failed",
                    "type": "string"
                },
                "user_id": {
                    "description": "Foydalanuvchi IDsi",
                    "type": "string"
                }
            }
        },
        "models.UnlockPDFRequest": {
            "type": "object",
            "required": [
                "input_file_id",
                "password"
            ],
            "properties": {
                "input_file_id": {
                    "description": "Kiruvchi PDF fayl IDsi",
                    "type": "string"
                },
                "password": {
                    "description": "PDF paroli",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}